move and update DYReLU
